requirements.txt
fastapi==0.103.2
pydantic==2.4.2
pydantic_core==2.10.1
pydantic-settings==2.0.3
pymongo==4.5.0
python-dotenv==1.0.0
starlette==0.27.0
uvicorn==0.23.2
